Re-add CF Workers destructuring workaround note
A user mentioned encountering the "This script will never generate a response" warning when running `vite dev` in a Vite+Typescript+Preact project for Cloudflare Workers. They were using wrangler 4.11.1, so the previously linked wrangler issue shouldn't be relevant (that one also didn't cause this warning if I'm not mistaken).

On a hunch, I asked them to try the classic destructuring workaround and it worked.

Not sure if this is an issue with vite, wrangler, Preact, or any other component in the mix in their project, but it seems this will be useful to people still.

(and the comment referring to the note was still in the code block anyway, so)